requirements.txt
requests
python-dotenv
google-search-results